:art: split screens in different files
